pyproject.toml
[project]
name = "kelen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "glcontext>=3.0.0",
    "moderngl>=5.12.0",
    "numpy>=2.4.2",
    "pygame-ce>=2.5.6",
]
